requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
requests>=2.26.0
aiofiles>=0.7.0
jinja2>=3.0.1
python-multipart>=0.0.5